Quick change for the VPC definition in virtual-private-cloud.adoc
Current VPC documentation refers to a VPC as a way to securely connect
to a customer’s private network.
This is really confusing for customers, since a VPC is basically just a
network without a connectivity model. They need a VPN tunnel, a VPC
peering connection or a DirectConnect for that.

It’s important to remark the difference between the VPC and the
connection to it.